型変換 cast の (double) が非推奨になったので、(float) に変更。 PHP Deprecated:  Non-canonical cast (double) is deprecated, use the (float) cast instead in /Users/yoya/git/yoya/IO_SWF/IO/SWF/Type/Double.php

+++ IO/SWF/Type/Double.php
@@ -15,10 +15,10 @@ class IO_SWF_Type_Double implements IO_SWF_Type {
     static function parse(&$reader, $opts = array()) {
         $data = $reader->getData(8);
 	$unpacked_data = unpack('d', $data);
-    	return (double)$unpacked_data[1];
+    	return (float)$unpacked_data[1];
     }
     static function build(&$writer, $value, $opts = array()) {
-        $data = pack('d', (double)$value);
+        $data = pack('d', (float)$value);
     	$writer->putData($data, 8);
     }
     static function string($value, $opts = array()) {
